[#119] AppConfigRepo 문서화 및 코드 정리
